voice_driver: don't use flite_phones_to_speech() for generating voice audio from phonemes
flite_phones_to_speech() needs a file to write the audio into. We used
the approach of transferring audio bytes to the caller by using a temporary
file that was given to flite_phones_to_speech() and then reading the
file contents of this file to the user provided buffer.

This was a cludge and doesn't work on Android, because there the temporary
directory is not well definied.
Use a direct approach for writing the voice audio to the user buffer instead

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>